requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-openai==0.0.5
langchain-core==0.1.10
langchain-community==0.0.10
langchain-chroma==0.0.5
langgraph==0.0.20
openai==1.3.7
chromadb==0.4.18
pypdf==3.17.4
serpapi==0.1.5
python-dotenv==1.0.0 